Add install and tests actions (#1)
* Add install and tests actions

* Update actions

* Add isort

* Change runner to matrix.os

* Fix yaml

* Update python versions

* Use fail-fast false

* Rename workflow

* Remove ~ in venv path

* Remove poetry install pip

* Factor with poetry action

* Remove needs key

* Move action into separate repo

* Fix action link

* Do not use action

* Add pytest cov

* Put virtualenvs in /Users/clementwalter

* Add coverage report

+++ tests/addons/assert_test.py
@@ -11,7 +11,7 @@ class TestAssert:
         @pytest.mark.parametrize("pandas_object", DEFAULT_PANDAS_OBJECTS)
         def test_should_raise_value_error_when_bool_fails(self, pandas_object):
             with pytest.raises(ValueError):
-                assert_(pandas_object([]), lambda x: x)
+                assert_(pandas_object([], dtype="float64"), lambda x: x)
 
         def test_should_raise_on_failure(self):
             with pytest.raises(AssertionError):
